[model] + [demo data] Added a wip LCA (GREET) script and LCA demo data
Initial LCA script which has to be further developed. Created a new folder for Cambium data which has to be populated when scenarios are run. Tied the LCA script with the outputs from RODeO. Saved files from LCA script are saved in the Output folder where RODeO outputs are saved.