Disable analysis tests
Met with Kaitlin Brunik (@kbrunik) today to discuss these particular analysis tests. These are based on outdated examples, and need to be re-assessed alongside the code in the holdover `to_organize` directory, which contains a number of files that likely belong as project code rather than usage examples for the repository.