Cleanup + test fixes (#189)
* Remove redundant resource files

* Fix clustering tests

* Remove unneeded tests

We determined in our last Refactor meeting that these are not real tests and should be removed.

* Limit shapely version

There's currently a range of shapely versions that work for HOPP, and this is reflected in `master`. If we want to loosen this constraint, we'll need to figure out why the tests fail on current versions of the package.

* Fix test_csp imports

* Revert PySAM to 3.0.0

This change broke tests in tests/hopp/test_hybrid, I'm not sure why, but we'll need to revisit if we want to upgrade.

* Fix documentation build

* Disable test_pressure_vessel

* Fix test_custom_financial

* Reorganize tests into tests/hopp

* Disable analysis tests

Met with Kaitlin Brunik (@kbrunik) today to discuss these particular analysis tests. These are based on outdated examples, and need to be re-assessed alongside the code in the holdover `to_organize` directory, which contains a number of files that likely belong as project code rather than usage examples for the repository.

* Update Custom Financial Model (#195)

* pull in changes from pysam_update_capacity

* fix import

* update tests

* undo a commenting

* BatteryStateless (#196)

* pull in changes from pysam_update_capacity

* fix import

* update tests

* add batttery_stateless

* break out create_max_gross_profit_objective by tech

* fix tests

* fix minor comments

* update .gitignore

* Remove more resource files

Another attempt at removing files that are not explicitly used in the code.

* Re-add clobbered changes and fix imports

---------

Co-authored-by: Darice L Guittet <dguittet@nrel.gov>
Co-authored-by: Jared Thomas <jaredthomas68@gmail.com>